refactor(frontend): 댓글 컴포넌트 공통 경로로 분리
커뮤니티 전용 댓글 컴포넌트를 (common)/comment로 이동해 재사용 가능한 구조로 정리하고, 상세 페이지의 참조 경로를 새 위치로 갱신했다.

Made-with: Cursor
