+++ frontend/src/app/(common)/comment/_components/CommentItem.tsx
@@ -174,4 +174,3 @@ export function CommentItem({
     </li>
   );
 }
-
